experiments/requirements.txt
requests
pandas[plot, output_formatting]
scikit-learn
numpy
bipartite_learn
scikit-posthocs
pyyaml
tqdm
